update the user routes

+++ BackEnd/app/routes/child_routes.py
@@ -1,5 +1,5 @@
 from flask_restx import Namespace, Resource
-from flask_jwt_extended import jwt_required, get_jwt_identity
+from flask_jwt_extended import jwt_required, get_jwt_identity,  get_jwt
 from marshmallow import ValidationError
 
 from app.services.child_service import ChildService
@@ -17,6 +17,12 @@ child_update_schema = ChildUpdateSchema()
 
 child_model, child_update_model = get_child_models(api)
 
+def require_parent():
+    claims = get_jwt()
+    if claims.get("role") != "parent":
+        return {"error": "Parent access required"}, 403
+    return None
+
 
 @api.route("/")
 class ChildListResource(Resource):
@@ -26,6 +32,9 @@ class ChildListResource(Resource):
     @api.expect(child_model, validate=True)
     def post(self):
         parent_id = get_jwt_identity()
+        error = require_parent()
+        if error:
+            return error
 
         try:
             child_data = child_create_schema.load(api.payload)
@@ -43,6 +52,9 @@ class ChildListResource(Resource):
     @api.doc(security="JWT")
     def get(self):
         parent_id = get_jwt_identity()
+        error = require_parent()
+        if error:
+            return error
         children = child_service.get_children_by_parent(parent_id)
 
         return children_response_schema.dump(children), 200
@@ -55,6 +67,9 @@ class ChildResource(Resource):
     @api.doc(security="JWT")
     def get(self, child_id):
         parent_id = get_jwt_identity()
+        error = require_parent()
+        if error:
+            return error
         child = child_service.get_child_for_parent(child_id, parent_id)
 
         if not child:
@@ -67,6 +82,9 @@ class ChildResource(Resource):
     @api.expect(child_update_model, validate=True)
     def put(self, child_id):
         parent_id = get_jwt_identity()
+        error = require_parent()
+        if error:
+            return error
 
         try:
             child_data = child_update_schema.load(api.payload)
@@ -90,6 +108,9 @@ class ChildResource(Resource):
     @api.doc(security="JWT")
     def delete(self, child_id):
         parent_id = get_jwt_identity()
+        error = require_parent()
+        if error:
+            return error
 
         deleted = child_service.delete_child_for_parent(child_id, parent_id)
